# app/delivery/web/user.go
package web

import (
	"github.com/gin-gonic/gin"
	"github.com/mateus-sousa-dev/meet-people/app/domain"
	"net/http"
)

type UserDelivery struct {
	useCase domain.UserUseCase
}

func NewUserDelivery(useCase domain.UserUseCase) *UserDelivery {
	return &UserDelivery{useCase: useCase}
}

// CreateUser godoc
// @Summary Create a user
// @Description Route to create a new user
// @Tags Users
// @Accept json
// @Produce json
// @Param user body domain.UserDto true "Modelo de usuário"
// @Success 201 {object} domain.User
// @Router /api/v1/users [post]
func (u *UserDelivery) CreateUser(c *gin.Context) {
	var userDto domain.UserDto
	err := c.ShouldBindJSON(&userDto)
	if err != nil {
		c.JSON(http.StatusBadRequest, err.Error())
		return
	}
	user, err := u.useCase.CreateUser(userDto)
	if err != nil {
		c.JSON(http.StatusInternalServerError, err.Error())
		return
	}
	c.JSON(http.StatusCreated, user)
}

// ActivateAccount godoc
// @Summary Activate Account
// @Description Route to Activate Account
// @Tags Users
// @Accept json
// @Produce json
// @Param activationpath query string true "Path de ativação"
// @Success 201 {string} account was activated successfully
// @Router /api/v1/activate-account/{activationpath} [get]
func (u *UserDelivery) ActivateAccount(c *gin.Context) {
	activationPath := c.Param("activationpath")
	err := u.useCase.ActivateAccount(activationPath)
	if err != nil {
		c.JSON(http.StatusInternalServerError, err.Error())
		return
	}
	c.JSON(http.StatusOK, "account was activated successfully")
}